Refactored evalate_rate_expression/evaluate_param_expression to go into kmos.utils
